Create amber-demand-charge.py
This script calculates the highest 30m interval. Within EMONCMS, when importing from the grid between 4pm and 9pm, only that data is posted as a seperate field, in this case it is looking up "SOLAREDGE/NET-DEMAND-IMPORT-KWH". from there, it calculates the highest 30m interval and multiplies it by the demand charge cost. This is specific to Energex in South East Queensland and their demand charges within this demand time window.

the script posts data for this month.

Then the script looks up previous months data, if it doesn't match what is already in the previous months data, then it posts the new data. this is meant to represent when a new month changes, since it is a new month then the previous month won't match the latest data, therefore it'll post the new data.

That way, the previous month post is only posted at the start of each month. Now each month demand charges can be totalled over the specific time period (i.e a year, two years etc.)