Create hotwater-solar.py
This script controls the heating of our 250L electric hotwater tank.

When there is excess solar, it activates the hotwater shelly1PM relay.

When there has not been any excess solar until 1PM AND the hotwater has consumed less than 16kwh, then it will activate the hotwater regardless from 1PM to 3PM. this is based on our hotwater usage, and temperature set to 75°C.

This script will only run from 8AM until 4PM. this is done to prevent heating the hotwater when our x3 10kWh batteries are discharging over night time AND it prevents the relay from the "ripple" effect during the early hours. the relay will "ripple" (turn off and on) quite often when it is just on the edge of enough solar excess.

There are dummy switches in the background, controlled from Apple HomeKit. We active the "Away mode" switch, this forces the hotwater relay to OFF and also does the same for our Airtouch4 Ducted air-conditioning.